Add basic IO Redirection
Adds a method to check for IO redirection operators in args list
and uses it to loop through args list, checking for operator and
setting input and output targets to the next item in args array.
If IO redirection is detected, it will pass the parameters into the
forkexec command to use freopen after forking in the child process.